+++ src/containers/MainSlider.js
