+++ src/containers/SliderBox.js
