+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDom from 'react-dom';
 import App from './components/App';
 import { Provider } from 'react-redux';
-import store from './store';
+import store from './store/store';
 import './reset.css';
 
 ReactDom.render(
